Move Angular components and tests to `src/sidebar/components`
This makes the dir name match the type of object these modules export.
