+++ src/sidebar/app.js
@@ -136,26 +136,26 @@ module.exports = angular.module('h', [
   .directive('hypothesisApp', require('./directive/app'))
 
   // UI components
-  .component('annotation', require('./directive/annotation').component)
-  .component('annotationShareDialog', require('./directive/annotation-share-dialog'))
-  .component('annotationThread', require('./directive/annotation-thread'))
-  .component('dropdownMenuBtn', require('./directive/dropdown-menu-btn'))
-  .component('helpLink', require('./directive/help-link'))
-  .component('helpPanel', require('./directive/help-panel'))
-  .component('loggedoutMessage', require('./directive/loggedout-message'))
-  .component('loginControl', require('./directive/login-control'))
-  .component('loginForm', require('./directive/login-form').component)
-  .component('publishAnnotationBtn', require('./directive/publish-annotation-btn'))
-  .component('searchInput', require('./directive/search-input'))
-  .component('searchStatusBar', require('./directive/search-status-bar'))
-  .component('selectionTabs', require('./directive/selection-tabs'))
-  .component('sidebarTutorial', require('./directive/sidebar-tutorial').component)
-  .component('shareDialog', require('./directive/share-dialog'))
-  .component('sortDropdown', require('./directive/sort-dropdown'))
-  .component('svgIcon', require('./directive/svg-icon'))
-  .component('tagEditor', require('./directive/tag-editor'))
-  .component('threadList', require('./directive/thread-list'))
-  .component('timestamp', require('./directive/timestamp'))
+  .component('annotation', require('./components/annotation').component)
+  .component('annotationShareDialog', require('./components/annotation-share-dialog'))
+  .component('annotationThread', require('./components/annotation-thread'))
+  .component('dropdownMenuBtn', require('./components/dropdown-menu-btn'))
+  .component('helpLink', require('./components/help-link'))
+  .component('helpPanel', require('./components/help-panel'))
+  .component('loggedoutMessage', require('./components/loggedout-message'))
+  .component('loginControl', require('./components/login-control'))
+  .component('loginForm', require('./components/login-form').component)
+  .component('publishAnnotationBtn', require('./components/publish-annotation-btn'))
+  .component('searchInput', require('./components/search-input'))
+  .component('searchStatusBar', require('./components/search-status-bar'))
+  .component('selectionTabs', require('./components/selection-tabs'))
+  .component('sidebarTutorial', require('./components/sidebar-tutorial').component)
+  .component('shareDialog', require('./components/share-dialog'))
+  .component('sortDropdown', require('./components/sort-dropdown'))
+  .component('svgIcon', require('./components/svg-icon'))
+  .component('tagEditor', require('./components/tag-editor'))
+  .component('threadList', require('./components/thread-list'))
+  .component('timestamp', require('./components/timestamp'))
 
   // These should use `component()` but will require some changes.
   .directive('groupList', require('./directive/group-list').directive)
