+++ src/sidebar/components/annotation-share-dialog.js
