+++ src/sidebar/components/annotation-thread.js
