+++ src/sidebar/components/annotation.js
