+++ src/sidebar/components/help-panel.js
