+++ src/sidebar/components/loggedout-message.js
