+++ src/sidebar/components/login-control.js
