+++ src/sidebar/components/publish-annotation-btn.js
