+++ src/sidebar/components/search-input.js
