+++ src/sidebar/components/search-status-bar.js
