+++ src/sidebar/components/selection-tabs.js
