+++ src/sidebar/components/share-dialog.js
