+++ src/sidebar/components/sidebar-tutorial.js
