+++ src/sidebar/components/sort-dropdown.js
