+++ src/sidebar/components/svg-icon.js
