+++ src/sidebar/components/tag-editor.js
