+++ src/sidebar/components/test/publish-annotation-btn-test.js
@@ -2,7 +2,7 @@
 
 var angular = require('angular');
 
-var util = require('./util');
+var util = require('../../directive/test/util');
 
 var fakeStorage = {};
 var fakeLocalStorage = {
