+++ src/sidebar/components/test/search-input-test.js
@@ -2,7 +2,7 @@
 
 var angular = require('angular');
 
-var util = require('./util');
+var util = require('../../directive/test/util');
 
 describe('searchInput', function () {
   var fakeHttp;
