+++ src/sidebar/components/test/search-status-bar-test.js
@@ -2,7 +2,7 @@
 
 var angular = require('angular');
 
-var util = require('./util');
+var util = require('../../directive/test/util');
 
 describe('searchStatusBar', function () {
   before(function () {
