+++ src/sidebar/components/test/share-dialog-test.js
@@ -2,7 +2,7 @@
 
 var angular = require('angular');
 
-var util = require('./util');
+var util = require('../../directive/test/util');
 
 describe('shareDialog', function () {
   var fakeAnnotationUI;
