+++ src/sidebar/components/test/sidebar-tutorial-test.js
