+++ src/sidebar/components/test/tag-editor-test.js
@@ -2,7 +2,7 @@
 
 var angular = require('angular');
 
-var util = require('./util');
+var util = require('../../directive/test/util');
 
 describe('tagEditor', function () {
   var fakeTags;
